Finish Website Layout
I wrapped up the banner and tailored it to feature a project.
I competed the next section for additional projects. This was done
using flexbox (same as the header and nav bar). I then added in a
testimonial section that really serves as a mission statement as
a learning  web developer looking for work opportunities. The final
thing done was a simple footer. These last two additionas were also
using flexbox.

I will now focus on making the links functional and
adding visual adjustments to the layout.